# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Emailibrium API
  version: 0.1.0
  description: |
    Vector-native email intelligence platform API.

    Emailibrium provides semantic search, email classification, ingestion
    orchestration, and inbox intelligence via a REST + SSE interface backed
    by vector embeddings and HNSW indexing.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Emailibrium Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /ingestion/status:
    get:
      tags: [ingestion]
      summary: Ingestion progress (SSE stream)
      description: |
        Server-Sent Events stream that emits `progress` events with
        `IngestionProgress` payloads. Optionally filter by job_id.

        The stream sends a `ping` keep-alive every 15 seconds.
      operationId: ingestionStatus
      parameters:
        - name: job_id
          in: query
          required: false
          schema:
            type: string
          description: Filter events to a specific ingestion job
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Each event has the format:
                  ```
                  event: progress
                  data: {"job_id":"...","total":100,"processed":50,...}
                  ```
  /ingestion/start:
    post:
      tags: [ingestion]
      summary: Start email ingestion
      description: |
        Start a new ingestion job for the specified account. Returns
        immediately with a job_id that can be used to track progress
        via the SSE status endpoint.
      operationId: startIngestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRequest"
            example:
              account_id: "gmail-user-1"
              full_sync: false
      responses:
        "200":
          description: Ingestion job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
              example:
                job_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "started"
                message: "Ingestion started for account gmail-user-1"
        "500":
          description: Failed to start ingestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ingestion/pause:
    post:
      tags: [ingestion]
      summary: Pause ingestion
      description: Pause a running ingestion job. The job can be resumed later.
      operationId: pauseIngestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PauseResumeRequest"
            example:
              job_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Ingestion job paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
              example:
                job_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "paused"
                message: "Ingestion job paused"
        "500":
          description: Failed to pause ingestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ingestion/resume:
    post:
      tags: [ingestion]
      summary: Resume ingestion
      description: Resume a previously paused ingestion job.
      operationId: resumeIngestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PauseResumeRequest"
            example:
              job_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Ingestion job resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
              example:
                job_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "resumed"
                message: "Ingestion job resumed"
        "500":
          description: Failed to resume ingestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Insights
  # ---------------------------------------------------------------------------
  /insights/subscriptions:
    get:
      tags: [insights]
      summary: Detected subscriptions
      description: |
        Analyse the email corpus to detect subscription patterns based on
        List-Unsubscribe headers, sender frequency, and content similarity.
      operationId: getSubscriptions
      responses:
        "200":
          description: List of detected subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionInsight"
              example:
                - sender: "newsletter@techblog.com"
                  sender_name: "TechBlog Weekly"
                  category: "Newsletter"
                  email_count: 52
                  recurrence_pattern: "weekly"
                  has_unsubscribe: true
                  first_seen: "2025-01-15T10:00:00Z"
                  last_seen: "2026-03-20T10:00:00Z"
                  estimated_read_rate: 0.35
        "500":
          description: Insight analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /insights/recurring:
    get:
      tags: [insights]
      summary: Recurring sender patterns
      description: |
        Identify senders who email at regular intervals and classify
        their communication patterns.
      operationId: getRecurringSenders
      responses:
        "200":
          description: List of recurring sender patterns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringSenderInsight"
              example:
                - sender: "reports@analytics.com"
                  sender_name: "Analytics Dashboard"
                  email_count: 90
                  recurrence_pattern: "daily"
                  avg_interval_days: 1.02
                  first_seen: "2025-12-01T08:00:00Z"
                  last_seen: "2026-03-22T08:00:00Z"
        "500":
          description: Insight analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /insights/report:
    get:
      tags: [insights]
      summary: Inbox intelligence report
      description: |
        Generate an aggregated inbox health report including total email
        count, category breakdown, top senders, subscription count,
        and recommended actions.
      operationId: getInboxReport
      responses:
        "200":
          description: Inbox intelligence report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxReport"
              example:
                total_emails: 12450
                categories:
                  Work: 4200
                  Newsletter: 3100
                  Marketing: 2800
                  Personal: 1500
                  Finance: 450
                  Uncategorized: 400
                top_senders:
                  - sender: "notifications@github.com"
                    count: 1200
                  - sender: "newsletter@techblog.com"
                    count: 520
                subscription_count: 45
                estimated_time_savings_minutes: 120
                generated_at: "2026-03-23T14:30:00Z"
        "500":
          description: Report generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

# ---------------------------------------------------------------------------
# Schemas
# ---------------------------------------------------------------------------
components:
  schemas:
...
    StartRequest:
      type: object
      properties:
        account_id:
          type: string
          description: Email account identifier (defaults to "default")
        full_sync:
          type: boolean
          description: Whether to re-sync all emails (default false)
    PauseResumeRequest:
      type: object
      required: [job_id]
      properties:
        job_id:
          type: string
          description: Ingestion job identifier
    JobResponse:
      type: object
      properties:
        job_id:
          type: string
          description: Unique job identifier (UUID)
        status:
          type: string
          enum: [started, paused, resumed]
          description: Current job status
        message:
          type: string
          description: Human-readable status message
...
    SubscriptionInsight:
      type: object
      properties:
        sender:
          type: string
          format: email
          description: Sender email address
        sender_name:
          type: string
          description: Display name of the sender
        category:
          type: string
          enum: [Newsletter, Marketing, Notification, Receipt, Social, Unknown]
          description: Subscription category
        email_count:
          type: integer
          format: int64
          description: Total emails received from this sender
        recurrence_pattern:
          type: string
          description: Detected send frequency
        has_unsubscribe:
          type: boolean
          description: Whether List-Unsubscribe header is present
        first_seen:
          type: string
          format: date-time
          description: Date of the earliest email from this sender
        last_seen:
          type: string
          format: date-time
          description: Date of the most recent email from this sender
        estimated_read_rate:
          type: number
          format: float
          description: Estimated percentage of emails opened (0.0 to 1.0)
    RecurringSenderInsight:
      type: object
      properties:
        sender:
          type: string
          format: email
        sender_name:
          type: string
        email_count:
          type: integer
          format: int64
        recurrence_pattern:
          type: string
          description: Detected pattern (daily, weekly, monthly, etc.)
        avg_interval_days:
          type: number
          format: float
          description: Average days between emails
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
    InboxReport:
      type: object
      properties:
        total_emails:
          type: integer
          format: int64
        categories:
          type: object
          additionalProperties:
            type: integer
            format: int64
          description: Email count per category
        top_senders:
          type: array
          items:
            type: object
            properties:
              sender:
                type: string
              count:
                type: integer
                format: int64
        subscription_count:
          type: integer
          description: Number of detected subscriptions
        estimated_time_savings_minutes:
          type: integer
          description: Estimated minutes saved by using Emailibrium
        generated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message